* config now only needs a media path and a plugin path * added command line parameter to force config dialog * adapted plugin and resource loading to use config paths * relocated resources.cfg

+++ RobotPlugin/RobotEntityView.cpp
@@ -8,8 +8,8 @@ RobotEntityView::RobotEntityView( Entity* Object, Ogre::SceneManager* Mgr )
 {
 	mNode = GetSceneManager()->getRootSceneNode()->createChildSceneNode();
 
-    //only if Model is set yet...
-    if (Object->GetAttribute("Model") != 0)	OnEntityAttributeChanged(Object,"Model",Object->GetAttribute("Model"));
+	//only if Model is set yet...
+	if (Object->GetAttribute("Model") != 0)	OnEntityAttributeChanged(Object,"Model",Object->GetAttribute("Model"));
 }
 
 RobotEntityView::~RobotEntityView()
@@ -54,7 +54,7 @@ void RobotEntityView::OnEntityAttributeChanged(
 			for ( iterator i=Skeleton->getBoneIterator(); i.hasMoreElements(); )
 			{
 				Ogre::Bone* Bone=i.getNext();
-                    GetDataEntity()->RemoveAttribute(Bone->getName());
+					GetDataEntity()->RemoveAttribute(Bone->getName());
 
 			}
 
@@ -89,12 +89,12 @@ void RobotEntityView::OnEntityAttributeChanged(
 			const quat Prefix=Bone->getInitialOrientation().Inverse();
 			const quat Delta=Prefix*Bone->getOrientation();
 
-            //if this attribute has already a attribute, do not overwrite it and inform observers (to  get visualize the orientation here)
-            if ( GetDataEntity()->GetAttribute(Name) == 0 ) {
-                GetDataEntity()->Set<double>(Name,Delta.getPitch().valueDegrees());
-            } else {
-                GetDataEntity()->Changed(Name, GetDataEntity()->GetAttribute(Name));
-            }
+			//if this attribute has already a attribute, do not overwrite it and inform observers (to  get visualize the orientation here)
+			if ( GetDataEntity()->GetAttribute(Name) == 0 ) {
+				GetDataEntity()->Set<double>(Name,Delta.getPitch().valueDegrees());
+			} else {
+				GetDataEntity()->Changed(Name, GetDataEntity()->GetAttribute(Name));
+			}
 		}
 	}
 	else if (Rhs->GetAttribute("Model") != 0)
@@ -104,16 +104,16 @@ void RobotEntityView::OnEntityAttributeChanged(
 
 		// Check if this is an angle value and a bone of this name exists
 		Ogre::SkeletonInstance* Skeleton=mOgreEntity->getSkeleton();
-		const double* Angle = boost::get<double>(Attribute);
+		const float* Angle = boost::get<float>( Attribute );
 
 		if ( Angle && Skeleton->hasBone(Name) )
 		{
 			Ogre::Bone* Bone=Skeleton->getBone(Name);
 
 			quat Delta;
-			Delta.FromAngleAxis(Ogre::Degree(*Angle),vec3(0.f,1.f,0.f));
+			Delta.FromAngleAxis( Ogre::Degree( *Angle ), vec3( 0.f, 1.f, 0.f ) );
 
-			Bone->setOrientation(Bone->getInitialOrientation()*Delta);
+			Bone->setOrientation( Bone->getInitialOrientation() * Delta );
 
 			// FIXME: is this really needed?
 			mOgreEntity->getAllAnimationStates()->_notifyDirty();
